Screen/PortableColor: fix return type of GetLuminosity()

+++ src/Screen/PortableColor.hpp
@@ -165,7 +165,7 @@ public:
   constexpr Luminosity8(uint8_t r, uint8_t g, uint8_t b)
     :value((r * 2126 + g * 7152 + b * 722 + 5000) / 10000) {}
 
-  constexpr uint16_t GetLuminosity() const {
+  constexpr uint8_t GetLuminosity() const {
     return value;
   }
 };
